Move monitoring API to API2
