+++ src/test/java/org/craftercms/studio/test/api2/objects/MonitoringAPI2.java
@@ -15,10 +15,11 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.craftercms.studio.test.api.objects;
+package org.craftercms.studio.test.api2.objects;
 
 
 import org.apache.http.HttpStatus;
+import org.craftercms.studio.test.api.objects.BaseAPI;
 import org.craftercms.studio.test.utils.APIConnectionManager;
 import org.craftercms.studio.test.utils.JsonTester;
 
@@ -26,23 +27,23 @@ import org.craftercms.studio.test.utils.JsonTester;
  * @author chris lim
  *
  */
-public class MonitoringAPI extends BaseAPI {
+public class MonitoringAPI2 extends BaseAPI {
 
 
-	public MonitoringAPI(JsonTester api, APIConnectionManager apiConnectionManager) {
+	public MonitoringAPI2(JsonTester api, APIConnectionManager apiConnectionManager) {
 		super(api, apiConnectionManager);
 	}
 
  	public void testVersion(){
-   		api.get("studio/api/1/services/api/1/monitor/version.json").execute().status(HttpStatus.SC_OK).debug();
+   		api.get("studio/api/2/monitoring/version.json").execute().status(HttpStatus.SC_OK).debug();
  	}
  	
  	public void testStatus(){
-   		api.get("studio/api/1/services/api/1/monitor/status.json").execute().status(HttpStatus.SC_OK).debug();
+   		api.get("studio/api/2/monitoring/status.json").execute().status(HttpStatus.SC_OK).debug();
  	}
 
  	public void testMemory(){
-   		api.get("studio/api/1/services/api/1/monitor/memory.json").execute().status(HttpStatus.SC_OK).debug();
+   		api.get("studio/api/2/monitoring/memory.json").execute().status(HttpStatus.SC_OK).debug();
  	}
  	
 }
